backend/requirements/base.txt

# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
coverage==5.2.1
django==3.0.9
django-cors-headers==3.5
django-countries==6.1.2
django-currencies==0.10.1
django-currentuser==0.5.0
django-filter==2.3.0
django-graphql-auth==0.3.11
django-graphql-jwt==0.3.0
django-money==1.1
django-richenum==3.8.0
graphene-django==2.13.0
gunicorn==20.0.4
pillow==7.2.0
pytest-django==3.9.0
python-dateutil==2.8.1
psycopg2-binary==2.8.5